pyproject.toml
[tool.poetry]
name = "minigest"
version = "0.1.0"
description = "Un mini gestionale pensato per aiutare a mantenere i conti di un'azienda"
authors = ["Davide Di Criscito <davide.dicriscito@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/dcdeiv/minigest"
keywords = ["ERP", "django", "gestionale"]
classifier = [
  "Environment :: Web Environment",
  "Framework :: Django",
  "Framework :: Django :: 3.0",
  "Intended Audience :: Developers",
  "Natural Language :: Italian",
  "Topic :: Office/Business"
]

[tool.black]
line-length = 88
exclude = '''
/(
    \.eggs
  | \.git
  | \.mypy_cache
  | \.tox
  | \.venv
  | dist
  | django_erp/migrations
)/
'''

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
tag_format = "$major.$minor.$patch$prerelease"
bump_message = "release $current_version → $new_version"
version_files = [
    "pyproject.toml:version"
]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.6"
django-polymorphic = "^2.1.2"

[tool.poetry.dev-dependencies]
commitizen = "^1.22.0"
flake8 = "^3.8.2"
flake8-bugbear = "^20.1.4"
black = "^19.10b0"
isort = "^4.3.21"
pre-commit = "^2.4.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
